% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blast_functions - Metascope Update.R
\name{blast_result_metrics}
\alias{blast_result_metrics}
\title{Blast Result Metrics}
\usage{
blast_result_metrics(
  blast_results_table_path,
  uniqueness_score_by = "species",
  percentage_hit_by = "species",
  contaminant_score_by = "genus"
)
}
\arguments{
\item{blast_results_table_path}{}
}
\value{
Loads blast result table (csv) and generates a list of best_hit, uniqueness_score,
percentage_hit and contaminant score

TODO: ADD ERROR HANDLING
}
\description{
Blast Result Metrics
}
