% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBlast_results.R
\name{get_seqs}
\alias{get_seqs}
\title{get_seqs}
\usage{
get_seqs(id, bam_file, n = 10)
}
\arguments{
\item{id}{Taxonomy ID of genome to get sequences from}

\item{bam_file}{A sorted bam file and index file, loaded with Rsamtools::bamFile}

\item{n}{Number of sequences to retrieve}
}
\value{
Biostrings format sequences
}
\description{
get_seqs
}
