% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_blast_results.R
\name{ncbi_blast_results}
\alias{ncbi_blast_results}
\title{Blast Results to NCBI API}
\usage{
ncbi_blast_results(
  results_table,
  bam_file,
  num_results = 10,
  num_reads_per_result = 10,
  hit_list = "10",
  out_path
)
}
\arguments{
\item{results_table}{A dataframe of the Metascope results}

\item{bam_file}{A sorted bam file and index file, loaded with Rsamtools::bamFile}

\item{num_results}{A number indicating number of Metascope results to blast}

\item{num_reads_per_result}{A number indicating number of reads to blast per result}

\item{hit_list}{A number of how many blast results to fetch for each read}

\item{out_path}{Output directory to save csv files, including basename of files}
}
\value{
Creates and exports num_results number of csv files with blast results from NCBI
}
\description{
Blast Results to NCBI API
}
