% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBlast_results.R
\name{rBlast_results}
\alias{rBlast_results}
\title{Local Blast Results}
\usage{
rBlast_results(
  results_table,
  bam_file,
  num_results = 10,
  num_reads_per_result = 100,
  hit_list = "10",
  db_path,
  out_path
)
}
\arguments{
\item{results_table}{A dataframe of the Metascope results}

\item{bam_file}{A sorted bam file and index file, loaded with Rsamtools::bamFile}

\item{num_results}{A number indicating number of Metascope results to blast}

\item{num_reads_per_result}{A number indicating number of reads to blast per result}

\item{hit_list}{A number of how many blast results to fetch for each read}

\item{db_path}{Blast database path}

\item{out_path}{Output directory to save csv files, including basename of files}
}
\value{
Creates and exports num_results number of csv files with blast results from local blast
}
\description{
Local Blast Results
}
