% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBlast_results.R
\name{rBLAST_single_result}
\alias{rBLAST_single_result}
\title{rBLAST_single_result}
\usage{
rBLAST_single_result(
  results_table,
  bam_file,
  which_result = 1,
  num_reads = 100,
  hit_list = 10,
  db_path,
  quiet = TRUE
)
}
\arguments{
\item{results_table}{A dataframe of the Metascope results}

\item{bam_file}{A sorted bam file and index file, loaded with Rsamtools::bamFile}

\item{which_result}{Index in results_table for which result to Blast search}

\item{num_reads}{Number of reads to blast per result}

\item{hit_list}{Number of how many blast results to fetch per read}

\item{db_path}{Blast database path}
}
\value{
Returns a dataframe of blast results for a metascope result
}
\description{
rBLAST_single_result
}
