% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{metascope_blast}
\alias{metascope_blast}
\title{MetaScope Blast}
\usage{
metascope_blast(
  metascope_id_path,
  tmp_dir,
  out_dir,
  sample_name,
  num_reads = 100,
  hit_list = "10",
  db_path,
  uniqueness_score_by = "species",
  percentage_hit_by = "species",
  contaminant_score_by = "genus"
)
}
\arguments{
\item{metascope_id_path}{}
}
\description{
MetaScope Blast
}
